Time: 3 ms (60.84%) | Memory: 10.6 MB (45.73%) - LeetSync

+++ LeetCode/1293-three-consecutive-odds/three-consecutive-odds.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     bool threeConsecutiveOdds(vector<int>& arr) {
-        short count = 0;
+        int count = 0;
         for (auto iter = arr.begin(); iter != arr.end(); iter++) {
             if (*iter % 2 == 1) {
                 count += 1;
